Move assertions to separate file from helpers.js.
